+++ src/components/Scroller/Scroller.mdx
@@ -1,4 +1,4 @@
-import { Meta, Canvas } from '@storybook/blocks';
+import { Meta } from '@storybook/blocks';
 
 import * as ScrollerStories from './Scroller.stories.svelte';
 
@@ -6,7 +6,9 @@ import * as ScrollerStories from './Scroller.stories.svelte';
 
 # Scroller
 
-The `Scroller` component TK
+The `Scroller` component TK.
+
+[Demo](?path=/story/components-graphics-scroller--demo)
 
 ```svelte
 <script>
@@ -15,5 +17,3 @@ The `Scroller` component TK
 
 <Scroller />
 ```
-
-<Canvas of={ScrollerStories.Test} />
